==> 8. Use Cases/fabric_benchmark_suite/00_setup.ipynb <==
# Setup Experiments

## Create asset folders

Let's create all the folders that are needed to store the intermediate results, raw results and analysis reports.

from common.config import create_experiment_folders

create_experiment_folders()
